requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
firebase-admin>=6.5.0
firebase-functions>=0.1.1
google-cloud-firestore>=2.20.0
pydantic>=2.11.7
python-multipart>=0.0.20
requests>=2.31.0
python-dotenv>=1.0.0
httpx>=0.27.0 